requirements.txt
Flask
configparser
pyserial
crcmod
pyowm
pytz
pyopenssl
twilio
chump
paho-mqtt
smbus
pysnmp
ldap3
pyasn1
pyotp==2.3.0
